svn merge -r1313:HEAD svn+ssh://dgc@rubyforge.org/var/svn/myexperiment/trunk
git-svn-id: svn+ssh://rubyforge.org/var/svn/myexperiment/branches/questions@1992 4f7b4649-9ca9-420a-81b3-55a188e4d46f

+++ app/controllers/blobs_controller.rb
@@ -4,36 +4,49 @@
 # See license.txt for details.
 
 class BlobsController < ApplicationController
-  before_filter :login_required, :except => [:index, :show, :download, :named_download, :search, :all]
+  before_filter :login_required, :except => [:index, :show, :download, :named_download, :statistics, :search, :all]
   
-  before_filter :find_blobs, :only => [:index, :all]
-  #before_filter :find_blob_auth, :only => [:download, :show, :edit, :update, :destroy]
+  before_filter :find_blobs, :only => [:all]
   before_filter :find_blob_auth, :except => [:search, :index, :new, :create, :all]
   
-  # GET /blobs;search
-  # GET /blobs.xml;search
+  before_filter :initiliase_empty_objects_for_new_pages, :only => [:new, :create]
+  before_filter :set_sharing_mode_variables, :only => [:show, :new, :create, :edit, :update]
+  
+  before_filter :check_is_owner, :only => [:edit, :update]
+  
+  # declare sweepers and which actions should invoke them
+  cache_sweeper :blob_sweeper, :only => [ :create, :update, :destroy ]
+  cache_sweeper :permission_sweeper, :only => [ :create, :update, :destroy ]
+  cache_sweeper :bookmark_sweeper, :only => [ :destroy, :favourite, :favourite_delete ]
+  cache_sweeper :tag_sweeper, :only => [ :create, :update, :tag, :destroy ]
+  cache_sweeper :download_viewing_sweeper, :only => [ :show, :download, :named_download ]
+  cache_sweeper :comment_sweeper, :only => [ :comment, :comment_delete ]
+  cache_sweeper :rating_sweeper, :only => [ :rate ]
+  
+  # GET /files;search
   def search
-
-    @query = params[:query] == nil ? "" : params[:query]
+    @query = params[:query] || ''
+    @query.strip!
     
-    @blobs = SOLR_ENABLE ? Blob.find_by_solr(@query, :limit => 100).results : []
+    @blobs = (SOLR_ENABLE && !@query.blank?) ? Blob.find_by_solr(@query, :limit => 100).results : []
     
     respond_to do |format|
       format.html # search.rhtml
-      format.xml  { render :xml => @blobs.to_xml }
     end
   end
   
-  # GET /blobs/1;download
+  # GET /files/1;download
   def download
-    @download = Download.create(:contribution => @blob.contribution, :user => (logged_in? ? current_user : nil))
+    if allow_statistics_logging(@blob)
+      @download = Download.create(:contribution => @blob.contribution, :user => (logged_in? ? current_user : nil), :user_agent => request.env['HTTP_USER_AGENT'], :accessed_from_site => accessed_from_website?())
+    end
     
-    send_data(@blob.data, :filename => @blob.local_name, :type => @blob.content_type)
+    send_data(@blob.content_blob.data, :filename => @blob.local_name, :type => @blob.content_type)
     
     #send_file("#{RAILS_ROOT}/#{controller_name}/#{@blob.contributor_type.downcase.pluralize}/#{@blob.contributor_id}/#{@blob.local_name}", :filename => @blob.local_name, :type => @blob.content_type)
   end
 
-  # GET /blobs/:id/download/:name
+  # GET /files/:id/download/:name
   def named_download
 
     # check that we got the right filename for this workflow
@@ -44,89 +57,89 @@ class BlobsController < ApplicationController
     end
   end
 
-  # GET /blobs
-  # GET /blobs.xml
+  # GET /files
   def index
     respond_to do |format|
       format.html # index.rhtml
-      format.xml  { render :xml => @blobs.to_xml }
     end
   end
   
-  # GET /blobs/all
+  # GET /files/all
   def all
     respond_to do |format|
       format.html # all.rhtml
     end
   end
   
-  # GET /blobs/1
-  # GET /blobs/1.xml
+  # GET /files/1
   def show
-    @viewing = Viewing.create(:contribution => @blob.contribution, :user => (logged_in? ? current_user : nil))
-    
-    @sharing_mode  = determine_sharing_mode(@blob)
-    @updating_mode = determine_updating_mode(@blob)
+    if allow_statistics_logging(@blob)
+      @viewing = Viewing.create(:contribution => @blob.contribution, :user => (logged_in? ? current_user : nil), :user_agent => request.env['HTTP_USER_AGENT'], :accessed_from_site => accessed_from_website?())
+    end
     
     respond_to do |format|
       format.html # show.rhtml
-      format.xml  { render :xml => @blob.to_xml }
     end
   end
   
-  # GET /blobs/new
+  # GET /files/new
   def new
-    @blob = Blob.new
-    
-    @sharing_mode  = 1
-    @updating_mode = 6
   end
   
-  # GET /blobs/1;edit
+  # GET /files/1;edit
   def edit
-    
-    @sharing_mode  = determine_sharing_mode(@blob)
-    @updating_mode = determine_updating_mode(@blob)
   end
   
   # POST /blobs
-  # POST /blobs.xml
   def create
     
-    params[:blob][:contributor_type], params[:blob][:contributor_id] = "User", current_user.id
-    params[:blob][:local_name] = params[:blob][:data].original_filename
-    params[:blob][:content_type] = params[:blob][:data].content_type
-    params[:blob][:data] = params[:blob][:data].read
-    
-    @blob = Blob.new(params[:blob])
-    
-    respond_to do |format|
-      if @blob.save
-        if params[:blob][:tag_list]
-          @blob.tags_user_id = current_user
-          @blob.tag_list = convert_tags_to_gem_format params[:blob][:tag_list]
-          @blob.update_tags
-        end
-        # update policy
-        @blob.contribution.update_attributes(params[:contribution])
-        
-        update_policy(@blob, params)
-        
-        update_credits(@blob, params)
-        update_attributions(@blob, params)
-        
-        flash[:notice] = 'File was successfully created.'
-        format.html { redirect_to file_url(@blob) }
-        format.xml  { head :created, :location => file_url(@blob) }
-      else
+    # don't create new blob if no file has been selected
+    if params[:blob][:data].size == 0
+      respond_to do |format|
+        flash.now[:error] = "Please select a file to upload."
         format.html { render :action => "new" }
-        format.xml  { render :xml => @blob.errors.to_xml }
+      end
+    else
+      data = params[:blob][:data].read
+      params[:blob][:local_name] = params[:blob][:data].original_filename
+      params[:blob][:content_type] = params[:blob][:data].content_type
+      params[:blob].delete('data')
+
+      params[:blob][:contributor_type], params[:blob][:contributor_id] = "User", current_user.id
+   
+      @blob = Blob.new(params[:blob])
+      @blob.content_blob = ContentBlob.new(:data => data)
+
+      respond_to do |format|
+        if @blob.save
+          if params[:blob][:tag_list]
+            @blob.tags_user_id = current_user
+            @blob.tag_list = convert_tags_to_gem_format params[:blob][:tag_list]
+            @blob.update_tags
+          end
+          # update policy
+          @blob.contribution.update_attributes(params[:contribution])
+        
+          policy_err_msg = update_policy(@blob, params)
+        
+          update_credits(@blob, params)
+          update_attributions(@blob, params)
+        
+          if policy_err_msg.blank?
+            flash[:notice] = 'File was successfully created.'
+            format.html { redirect_to file_url(@blob) }
+          else
+            flash[:notice] = "File was successfully created. However some problems occurred, please see these below.</br></br><span style='color: red;'>" + policy_err_msg + "</span>"
+            format.html { redirect_to :controller => 'blobs', :id => @blob, :action => "edit" }
+          end
+        else
+          format.html { render :action => "new" }
+        end
       end
     end
   end
   
-  # PUT /blobs/1
-  # PUT /blobs/1.xml
+  # PUT /files/1
   def update
     # hack for select contributor form
     if params[:contributor_pair]
@@ -146,23 +159,26 @@ class BlobsController < ApplicationController
     
     respond_to do |format|
       if @blob.update_attributes(params[:blob])
-        refresh_tags(@blob, params[:blob][:tag_list], current_user) if params[:blob][:tag_list]
-        update_policy(@blob, params)
+        @blob.refresh_tags(convert_tags_to_gem_format(params[:blob][:tag_list]), current_user) if params[:blob][:tag_list]
+        
+        policy_err_msg = update_policy(@blob, params)
         update_credits(@blob, params)
         update_attributions(@blob, params)
         
-        flash[:notice] = 'File was successfully updated.'
-        format.html { redirect_to file_url(@blob) }
-        format.xml  { head :ok }
+        if policy_err_msg.blank?
+          flash[:notice] = 'File was successfully updated.'
+          format.html { redirect_to file_url(@blob) }
+        else
+          flash[:error] = policy_err_msg
+          format.html { redirect_to :controller => 'blobs', :id => @blob, :action => "edit" }
+        end
       else
         format.html { render :action => "edit" }
-        format.xml  { render :xml => @blob.errors.to_xml }
       end
     end
   end
   
-  # DELETE /blobs/1
-  # DELETE /blobs/1.xml
+  # DELETE /files/1
   def destroy
     success = @blob.destroy
 
@@ -170,17 +186,14 @@ class BlobsController < ApplicationController
       if success
         flash[:notice] = "File has been deleted."
         format.html { redirect_to files_url }
-        format.xml  { head :ok }
       else
         flash[:error] = "Failed to delete File. Please contact your administrator."
         format.html { redirect_to file_url(@blob) }
-        format.xml  { render :xml => @blob.errors.to_xml }
       end
     end
   end
   
-  # POST /blobs/1;comment
-  # POST /blobs/1.xml;comment
+  # POST /files/1;comment
   def comment 
     text = params[:comment][:comment]
     
@@ -191,12 +204,10 @@ class BlobsController < ApplicationController
     
     respond_to do |format|
       format.html { render :partial => "comments/comments", :locals => { :commentable => @blob } }
-      format.xml { render :xml => @blob.comments.to_xml }
     end
   end
   
-  # DELETE /blobs/1;comment_delete
-  # DELETE /blobs/1.xml;comment_delete
+  # DELETE /files/1;comment_delete
   def comment_delete
     if params[:comment_id]
       comment = Comment.find(params[:comment_id].to_i)
@@ -208,53 +219,81 @@ class BlobsController < ApplicationController
     
     respond_to do |format|
       format.html { render :partial => "comments/comments", :locals => { :commentable => @blob } }
-      format.xml { render :xml => @blob.comments.to_xml }
     end
   end
   
-  # POST /blobs/1;rate
-  # POST /blobs/1.xml;rate
+  # POST /files/1;rate
   def rate
-    Rating.delete_all(["rateable_type = ? AND rateable_id = ? AND user_id = ?", @blob.class.to_s, @blob.id, current_user.id])
-    
-    @blob.ratings << Rating.create(:user => current_user, :rating => params[:rating])
-    
-    respond_to do |format|
-      format.html { 
-        render :update do |page|
-          page.replace_html "ratings_inner", :partial => "contributions/ratings_box_inner", :locals => { :contributable => @blob, :controller_name => controller.controller_name }
-          page.replace_html "ratings_breakdown", :partial => "contributions/ratings_box_breakdown", :locals => { :contributable => @blob }
-        end }
-      format.xml { render :xml => @rateable.ratings.to_xml }
+    if @blob.contributor_type == 'User' and @blob.contributor_id == current_user.id
+      error("You cannot rate your own file!", "")
+    else
+      Rating.delete_all(["rateable_type = ? AND rateable_id = ? AND user_id = ?", @blob.class.to_s, @blob.id, current_user.id])
+      
+      @blob.ratings << Rating.create(:user => current_user, :rating => params[:rating])
+      
+      respond_to do |format|
+        format.html { 
+          render :update do |page|
+            page.replace_html "ratings_inner", :partial => "contributions/ratings_box_inner", :locals => { :contributable => @blob, :controller_name => controller.controller_name }
+            page.replace_html "ratings_breakdown", :partial => "contributions/ratings_box_breakdown", :locals => { :contributable => @blob }
+          end }
+      end
     end
   end
   
-  # POST /blobs/1;tag
-  # POST /blobs/1.xml;tag
+  # POST /files/1;tag
   def tag
     @blob.tags_user_id = current_user # acts_as_taggable_redux
     @blob.tag_list = "#{@blob.tag_list}, #{convert_tags_to_gem_format params[:tag_list]}" if params[:tag_list]
     @blob.update_tags # hack to get around acts_as_versioned
     
-    expire_fragment(:controller => 'files', :action => 'all_tags')
-    expire_fragment(:controller => 'sidebar_cache', :action => 'tags', :part => 'most_popular_tags')
+    respond_to do |format|
+      format.html { 
+        render :update do |page|
+          unique_tag_count = @blob.tags.uniq.length
+          page.replace_html "mini_nav_tag_link", "(#{unique_tag_count})"
+          page.replace_html "tags_box_header_tag_count_span", "(#{unique_tag_count})"
+          page.replace_html "tags_inner_box", :partial => "tags/tags_box_inner", :locals => { :taggable => @blob, :owner_id => @blob.contributor_id } 
+        end
+      }
+    end
+  end
+  
+  # POST /files/1;favourite
+  def favourite
+    @blob.bookmarks << Bookmark.create(:user => current_user) unless @blob.bookmarked_by_user?(current_user)
     
     respond_to do |format|
-      format.html { render :partial => "tags/tags_box_inner", :locals => { :taggable => @blob, :owner_id => @blob.contributor_id } }
-      format.xml { render :xml => @blob.tags.to_xml }
+      flash[:notice] = "You have successfully added this item to your favourites."
+      format.html { redirect_to file_url(@blob) }
+    end
+  end
+  
+  # DELETE /files/1;favourite_delete
+  def favourite_delete
+    @blob.bookmarks.each do |b|
+      if b.user_id == current_user.id
+        b.destroy
+      end
+    end
+    
+    respond_to do |format|
+      flash[:notice] = "You have successfully removed this item from your favourites."
+      redirect_url = params[:return_to] ? params[:return_to] : file_url(@blob)
+      format.html { redirect_to redirect_url }
     end
   end
   
   protected
   
   def find_blobs
-   found = Blob.find(:all, 
+    found = Blob.find(:all, 
                        :order => "content_type ASC, local_name ASC, created_at DESC",
                        :page => { :size => 20, 
                        :current => params[:page] })
-      
+    
     found.each do |blob|
-      blob.data = nil unless blob.authorized?("download", (logged_in? ? current_user : nil))
+      blob.content_blob.data = nil unless blob.authorized?("download", (logged_in? ? current_user : nil))
     end
     
     @blobs = found
@@ -267,36 +306,63 @@ class BlobsController < ApplicationController
       if blob.authorized?(action_name, (logged_in? ? current_user : nil))
         @blob = blob
         
-        @blob_url = url_for :only_path => false,
+        @blob_entry_url = url_for :only_path => false,
                             :host => base_host,
                             :id => @blob.id
 
-        @named_download_url = url_for :action => 'named_download',
+        @named_download_url = url_for :controller => 'files',
+                                      :action => 'named_download',
                                       :id => @blob.id, 
-                                      :version => 1, # blobs aren't versioned (yet)
                                       :name => @blob.local_name
 
       else
         if logged_in? 
           error("File not found (id not authorized)", "is invalid (not authorized)")
+          return false
         else
           find_blob_auth if login_required
         end
       end
-      rescue ActiveRecord::RecordNotFound
+    rescue ActiveRecord::RecordNotFound
       error("File not found", "is invalid")
+      return false
+    end
+  end
+  
+  def initiliase_empty_objects_for_new_pages
+    if ["new", "create"].include?(action_name)
+      @blob = Blob.new
+    end
+  end
+  
+  def set_sharing_mode_variables
+    case action_name
+      when "new"
+        @sharing_mode  = 0
+        @updating_mode = 6
+      when "create", "update"
+        @sharing_mode  = params[:sharing][:class_id].to_i if params[:sharing]
+        @updating_mode = params[:updating][:class_id].to_i if params[:updating]
+      when "show", "edit"
+        @sharing_mode  = determine_sharing_mode(@blob)
+        @updating_mode = determine_updating_mode(@blob)
+    end
+  end
+  
+  def check_is_owner
+    if @blob
+      error("You are not authorised to manage this File", "") unless @blob.owner?(current_user)
     end
   end
   
   private
   
   def error(notice, message, attr=:id)
-    flash[:notice] = notice
+    flash[:error] = notice
      (err = Blob.new.errors).add(attr, message)
     
     respond_to do |format|
       format.html { redirect_to files_url }
-      format.xml { render :xml => err.to_xml }
     end
   end
 end
